Fix serial port init
Zigbee-shepherd starts cc-znp twice when starting, once as primary
initialization, and then another time when performing a software reset.
This caused the serial port to open twice, the second time the serial
port was already open en will therefore be closed before opening it
again, this causes a ‘close’ event which breaks the cc-znp
initialization process.